修复了app为None的bug，并将处理逻辑分开
